First dump of something...
Simple interpreter of a few instructions.   I need to change this design
from this commit so that it can have an interpreter and a compiler.  I may
use generics with traits for this?  I also need to introduce Operands and
not assume all values will be usize values.  As it stands this is direct
addressing only and I cannot differentiate from a literal number and a
direct address.