Scale back my ambitions on modelling this project (for now).  What I really wanted was to have n potential backends via a common build method.  That common build method would use a specific trait (e.g. Interpret, X86Compile, etc...). Each instruction would implement those traits allowing all implementations to be highly segregated.
My discovery of impl trait params I thought was a solution until I
hit the next snag.  All instructions are different impls and so Rust sees them
as Unsized.  My Rust brain is incapable of making this a DST solution.

For now interpret and compile will just be functions of Instruction.  It is
good enough for my project goals and perhaps after I learn a lot more Rust I
will rewrite this thing.